requirements-dev.txt
-r requirements.txt

beautifulsoup4
coverage
factory-boy
pytest-cov
pytest-django
pytest